- app version = 5.1.0 - implemented sub-components for search.vue blocks - refactored search.vue to separate flows using sub-components (STILL WIP) - fixed tab title size on small screens - added search variables to store - created search-mixin.ts for common getters, setters, etc. - misc cleanup
