+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "name-request",
-  "version": "5.0.44",
+  "version": "5.1.0",
   "private": true,
   "appName": "Name Request UI",
   "sbcName": "SBC Common Components",
