+++ src/components/new-request/search-components/corp-number-checkbox.vue
@@ -1,0 +1,52 @@
+<template>
+  <div id="corp-number-checkbox">
+    <v-tooltip
+      top min-width="390"
+      content-class="top-tooltip"
+      transition="fade-transition"
+      :disabled="isMobile"
+    >
+      <template v-slot:activator="{ on }">
+        <v-checkbox
+          v-model="noCorpNum"
+          id="corp-num-checkbox"
+          class="copy-small mt-0 pt-0"
+          hide-details
+          v-slot:label
+          v-on="on"
+        >
+          <span v-on="on" class="copy-small">I don't have a corporate number</span>
+        </v-checkbox>
+      </template>
+
+      <p>
+        If you don't have or don't know the corporation number of the business, enter the full legal name
+        of the business in its home jurisdiction.
+      </p>
+
+      <p>
+        Note: If the home jurisdiction requires a name reservation, you may want to complete a name search
+        in the home jurisdiction first to ensure that the name is available and then return to BC.
+      </p>
+    </v-tooltip>
+  </div>
+</template>
+
+<script lang="ts">
+import { Component, Mixins } from 'vue-property-decorator'
+import NestedSelect from '@/components/common/nested-select.vue'
+import { SearchMixin } from '@/mixins'
+
+@Component({
+  components: { NestedSelect }
+})
+export default class CorpNumberCheckbox extends Mixins(SearchMixin) {
+  get noCorpNum (): boolean {
+    return this.getHasNoCorpNum
+  }
+
+  set noCorpNum (value: boolean) {
+    this.setNoCorpNum(value)
+  }
+}
+</script>
