+++ src/components/new-request/search.vue
@@ -2,306 +2,133 @@
   <v-container fluid id="search-container" class="copy-normal pt-10 px-10"
     :class="(isFederal && isRestoration) ? 'pb-0' : 'pb-12'"
   >
-    <v-row>
-      <v-col cols="12" class="font-weight-bold h6">
-        Get started by selecting an action:
-      </v-col>
+    <header class="h6">Get started by selecting an action:</header>
 
+    <v-row class="pt-6">
       <!-- Request Action -->
-      <v-col cols="12" md="6">
-        <v-tooltip
-          top
-          content-class="top-tooltip"
-          transition="fade-transition"
-          :disabled="!showRequestActionTooltip || isMobile"
-        >
-          <template v-slot:activator="scope">
-            <div v-on="scope.on">
-              <NestedSelect
-                label="Action"
-                :menuItems="RequestActions"
-                :errorMessages="getErrors.includes('request_action_cd') ? 'Please select an action' : ''"
-                :value="request"
-                maxHeight="333"
-                @change="setClearErrors(); onRequestActionChange($event)"
-              />
-            </div>
-          </template>
-          <span>{{ request && request.text }}</span>
-        </v-tooltip>
-      </v-col>
+      <RequestAction />
 
       <!-- display a dummy input box when Request Action is not yet selected -->
-      <v-col v-if="!request" cols="12" md="6">
-        <v-text-field filled disabled hide-details label="Select an action first" />
-      </v-col>
+      <DummyInputBox v-if="!getSearchRequest" />
 
-      <!-- Business Lookup/Fetch -->
-      <v-col v-if="showBusinessLookup" cols="12" md="6" class="business-lookup">
-        <template v-if="!business">
-          <BusinessLookup
-            v-if="isAuthenticated"
-            :searchStatus="lookupActiveOrHistorical"
-            @business="onBusiness($event)"
-          />
-          <BusinessFetch
-            v-else
-            @business="onBusiness($event)"
-          />
-        </template>
-        <v-form v-else ref="businessLookupTextFieldForm">
-          <v-text-field
-            append-icon="mdi-close"
-            readonly
-            filled
-            :label="businessLookupLabel"
-            :value="business.legalName"
-            :rules="rules"
-            :hint="businessLookupHint"
-            @click:append="onBusiness(null)"
-            @keyup.delete="onBusiness(null)"
-          />
-        </v-form>
-      </v-col>
+      <!-- New BC Business flow -->
+      <template v-if="isNewBcBusiness">
+        <EntityType v-if="getLocation" />
+        <CompanyType v-if="entity_type_cd" />
+        <NumberedCompanyBullets v-if="isNumberedCompany" />
 
-      <!-- Jurisdiction -->
-      <v-col v-if="showJurisdiction" cols="12" :md="isSelectedCompanyXPro ? 4 : 6">
-        <NestedSelect
-          label="Select your home jurisdiction"
-          :menuItems="jurisdictionOptions"
-          :error-messages="getErrors.includes('jurisdiction') ? 'Please select a jurisdiction' : ''"
-          :value="jurisdiction"
-          @change="setClearErrors(); onJurisdictionChange($event)"
-        />
-      </v-col>
-
-      <!-- Entity Type -->
-      <v-col v-if="showEntityType" cols="12" :md="isXproFlow ? 4 : ( isConversion ? 12 : 6)">
-        <v-text-field
-          v-if="isConversion && isBenBusiness"
-          filled
-          disabled
-          hide-details
-          label="Benefit Company to Limited Company"
-        />
-        <v-tooltip
-          v-else
-          top
-          content-class="top-tooltip"
-          :disabled="getRequestActionCd !== NrRequestActionCodes.CONVERSION || !entityConversionText || isMobile"
-          transition="fade-transition"
-        >
-          <template v-slot:activator="scope">
-            <div v-on="scope.on">
-              <v-select
-                id="entity-type-options-select"
-                :label="isConversion ? 'Select type of business to alter into' : 'Select type of business in B.C.'"
-                :error-messages="getErrors.includes('entity_type_cd') ? 'Please select a business type' : ''"
-                :items="entityTypeOptions"
-                :menu-props="{ bottom: true, offsetY: true}"
-                ref="selectBusinessTypeRef"
-                hide-details="auto"
-                filled
-                :value="isConversion ? getOriginEntityTypeCd : entity_type_cd"
-                @change="setClearErrors(); entity_type_cd = $event"
-              >
-                <template v-slot:item="{ item }">
-                  <v-tooltip
-                    :right="isScreenLg"
-                    :left="!isScreenLg"
-                    :disabled="!item.blurbs || isMobile"
-                    :content-class="!isScreenLg ? 'left-tooltip' : ''"
-                    transition="fade-transition"
-                  >
-                    <template v-slot:activator="scope">
-                      <span
-                        v-on="scope.on"
-                        class="list-item"
-                        :class="{ 'last-select-item': item.value === 'INFO' }"
-                      >{{ item.text }}</span>
-                    </template>
-
-                    <div v-for="(blurb, index) in entityBlurbs(item.value)" :key="`blurb-${index}`">
-                      <span :class="{ 'tooltip-bullet': index !== 0}">
-                        {{ blurb }}
-                      </span>
-                    </div>
-                  </v-tooltip>
-                </template>
-              </v-select>
-            </div>
-          </template>
-          <span>{{ entityConversionText }}</span>
-        </v-tooltip>
-      </v-col>
-
-      <!-- once an entity type is selected, is Federal, or is Restorable, ... -->
-      <template v-if="entity_type_cd || isFederal || isRestorable">
-        <!-- Company Type -->
-        <v-col v-if="showCompanyTypeRadioButtons" cols="12">
-          <p class="font-weight-bold h6">Select a company type:</p>
-          <v-radio-group
-            v-model="selectedCompanyType"
-            class="mt-3"
-            hide-details
-            mandatory
-            row
-            @change="selectedCompanyType = $event"
-          >
-            <v-radio
-              id="named-company-radio"
-              label="Named Company"
-              :value="CompanyTypes.NAMED_COMPANY"
-            />
-            <v-radio
-              id="numbered-company-radio"
-              label="Numbered Company"
-              :value="CompanyTypes.NUMBERED_COMPANY"
-            />
-          </v-radio-group>
-        </v-col>
-
-        <!-- Numbered company (or amalgamation) -->
-        <template v-if="isNamedCompany || isAmalgamation">
-          <!-- Xpro/Federal bullets -->
-          <v-col v-if="isXproFlow && isFederal" cols="12" :md="isSelectedCompanyXPro ? 10 : 8">
-            <ul class="bullet-points">
-              <li>Federally incorporated businesses do not need a Name Request.</li>
-
-              <li v-if="isAmalgamation">
-                To register your extraprovincial amalgamation, download and complete
-                <a :href="federalAmalgamationFormLink">
-                  this form <v-icon small class="ml-1" color="primary">mdi-open-in-new</v-icon>
-                </a>.
-              </li>
-
-              <li v-else-if="isRestoration && isSelectedXproAndRestorable">
-                To reinstate your business, complete
-                <a :href="fullReinstatementFormLink">
-                  this form <v-icon small class="ml-1" color="primary">mdi-open-in-new</v-icon>
-                </a> for a full reinstatement or
-                <a :href="limitedReinstatementFormLink">
-                  this form  <v-icon small class="ml-1" color="primary">mdi-open-in-new</v-icon>
-                </a> for a limited reinstatement.
-              </li>
-
-              <li v-else>
-                You may register your extraprovincial business immediately using its existing name
-                at Corporate Online.
-              </li>
-            </ul>
-          </v-col>
-
-          <!-- XPRO/MRAS number/name search/input -->
-          <v-col v-if="!isFederal" cols="12" :md="(isXproFlow || showDesignation) ? 8 : 12">
+        <template v-if="isNamedCompany">
+          <v-col cols="12" :md="showDesignation ? '8' : '12'">
             <NameInput
-              id="name-input-component"
-              :is-mras-search="(isXproFlow && isMrasJurisdiction && !noCorpNum)"
+              :is-mras-search="(isXproFlow && isMrasJurisdiction && !getHasNoCorpNum)"
               @emit-corp-num-validity="corpNumValid = $event"
             />
           </v-col>
+          <Designation cols="12" md="4" />
+        </template>
+      </template>
 
-          <!-- Designation -->
-          <v-col v-if="showDesignation" cols="12" md="4">
-            <v-select
-              :error-messages="getErrors.includes('designation') ? 'Please enter a designation' : ''"
-              filled
-              :items="designationOptions"
-              label="Select a Designation"
-              :readonly="!entity_type_cd"
-              :menu-props="{ bottom: true, offsetY: true}"
-              :value="getDesignation"
-              @input="setDesignation($event)"
-              class="mb-n3"
-              @change="setClearErrors()"
-            />
-          </v-col>
+      <!-- New Xpro Business flow -->
+      <template v-else-if="isNewXproBusiness">
+        <EntityType v-if="getLocation && !isFederal" cols="12" md="4" />
+        <Jurisdiction />
+        <CompanyType v-if="showCompanyTypeRadioButtons" />
+      </template>
 
-          <!-- Corporate Number checkbox, only for Canadian MRAS jurisdictions -->
-          <v-col v-if="isCanadian && isMrasJurisdiction" class="d-flex justify-end">
-            <v-tooltip
-              top min-width="390"
-              content-class="top-tooltip"
-              transition="fade-transition"
-              :disabled="isMobile"
-            >
-              <template v-slot:activator="{ on }">
-                <v-checkbox
-                  v-model="noCorpNum"
-                  id="corp-num-checkbox"
-                  class="copy-small mt-0 pt-0"
-                  hide-details
-                  v-slot:label
-                  v-on="on"
-                >
-                  <template>
-                    <span v-on="on" class="copy-small">I don't have a corporate number</span>
-                  </template>
-                </v-checkbox>
-              </template>
+      <!-- Continuation In flow -->
+      <template v-else-if="isContinuationIn">
+        <EntityType v-if="getLocation" />
+        <CompanyType v-if="showCompanyTypeRadioButtons" />
+      </template>
 
-              <p>If you don't have or don't know the corporation number of the business, enter the full legal name
-                of the business in its home jurisdiction.</p>
+      <!-- Change Name flow -->
+      <template v-else-if="isChangeName">
+        <BusinessLookupFetch />
+        <Jurisdiction v-if="isChangeNameXpro" />
+        <CompanyType v-if="showCompanyTypeRadioButtons" />
+      </template>
 
-              <p>Note: If the home jurisdiction requires a name reservation, you may want to complete a name search
-                in the home jurisdiction first to ensure that the name is available and then return to BC.</p>
-            </v-tooltip>
-          </v-col>
+      <!-- Amalgamation flow -->
+      <template v-else-if="isAmalgamation">
+        <EntityType v-if="getLocation && !isFederal" />
+
+        <template v-if="isXproEntityType(getEntityTypeCd)">
+          <Jurisdiction />
         </template>
 
-        <!-- Numbered company (or Federal) -->
-        <template v-if="isNumberedCompany || isFederal" cols="12">
-          <v-col v-if="isConversion && !isAlterOnline(getConversionType)">
-            <div class="contact-registries">
-              <p>To complete this alteration, please contact us at:</p>
-              <p>
-                  <v-icon small>mdi-phone</v-icon>&nbsp;Canada and U.S. Toll Free:
-                  <a href="tel:+1877-370-1033">1-877-370-1033</a>
-              </p>
-              <p><v-icon small>mdi-phone</v-icon>&nbsp;Victoria Office:
-                <a href="tel:250-370-1033">250-370-1033</a>
-              </p>
-              <p><v-icon small>mdi-email</v-icon>&nbsp;Email:
-                <a href="mailto:BCRegistries@gov.bc.ca">BCRegistries@gov.bc.ca</a>
-              </p>
-            </div>
-          </v-col>
-          <v-col v-else-if="!isXproFlow" cols="12">
-            <ul class="bullet-points">
-              <li>Your business name will be the Incorporation Number assigned by the Registry.</li>
-              <li>You can change your business name at a later date.</li>
-              <li>It is not possible to request a specific Incorporation Number.</li>
-            </ul>
-          </v-col>
-
-          <!-- Go to COLIN / Incorporate Now buttons -->
-          <v-col v-if = "!isConversion || isAlterOnline(getConversionType)" cols="12"
-            class="d-flex justify-center"
-          >
-            <v-btn
-              v-if="showColinButton"
-              class="px-9"
-              id="go-to-colin-button"
-              :href="colinLink"
-              target="_blank"
-            >
-              Go to Corporate Online to {{ isConversion ? 'Alter' : 'Register' }}
-              <v-icon small class="ml-1">mdi-open-in-new</v-icon>
-            </v-btn>
-            <v-btn
-              v-else-if="showActionButton"
-              class="px-9"
-              id="incorporate-now-button"
-              @click="actionNowClicked()"
-            >
-              {{ actionNowButtonText }}
-            </v-btn>
-          </v-col>
+        <template v-else>
+          <CompanyType v-if="showCompanyTypeRadioButtons" />
         </template>
+      </template>
+
+      <!-- Conversion (aka Alteration) flow -->
+      <template v-else-if="isConversion">
+        <EntityType v-if="getSearchBusiness && isAlterable" />
+        <BusinessLookupFetch />
+        <CompanyType v-if="showCompanyTypeRadioButtons" />
+      </template>
+
+      <!-- Restoration / Reinstatement flow -->
+      <template v-else-if="isRestoration">
+        <EntityType cols="12" md="6" />
+        <BusinessLookupFetch />
+        <Jurisdiction v-if="isSelectedXproAndRestorable" cols="12" md="4" />
+        <CompanyType v-if="showCompanyTypeRadioButtons" />
       </template>
     </v-row>
 
+    <!-- Xpro/Federal bullets -->
+    <!-- *** TODO: move this into flows above -->
+    <!-- *** TODO: simply this button's logic -->
+    <template v-if="entity_type_cd || isFederal || isRestorable">
+      <v-row v-if="isNamedCompany || isAmalgamation">
+        <XproFederalBullets v-if="isXproFlow && isFederal" cols="12" />
+      </v-row>
+    </template>
+
+    <!-- Corporate Number checkbox, only for Canadian MRAS jurisdictions -->
+    <!-- *** TODO: move this into flows above -->
+    <!-- *** TODO: simply this button's logic -->
+    <template v-if="entity_type_cd || isFederal || isRestorable">
+      <v-row v-if="isNamedCompany || isAmalgamation">
+        <v-col cols="12" class="d-flex justify-end">
+          <CorpNumberCheckbox v-if="isCanadian && isMrasJurisdiction" />
+        </v-col>
+      </v-row>
+    </template>
+
+    <!-- Go to COLIN / Incorporate Now buttons -->
+    <!-- *** TODO: keep this as common code at bottom of template -->
+    <!-- *** TODO: simply this button's logic -->
+    <template v-if="entity_type_cd || isFederal || isRestorable">
+      <v-row v-if="isNumberedCompany || isFederal">
+        <v-col v-if = "!isConversion || isAlterOnline(getConversionType)" cols="12"
+          class="d-flex justify-center"
+        >
+          <v-btn
+            v-if="showColinButton"
+            class="px-9"
+            id="go-to-colin-button"
+            :href="colinLink"
+            target="_blank"
+          >
+            Go to Corporate Online to {{ isConversion ? 'Alter' : 'Register' }}
+            <v-icon small class="ml-1">mdi-open-in-new</v-icon>
+          </v-btn>
+          <v-btn
+            v-else-if="showActionButton"
+            class="px-9"
+            id="incorporate-now-button"
+            @click="actionNowClicked()"
+          >
+            {{ actionNowButtonText }}
+          </v-btn>
+        </v-col>
+      </v-row>
+    </template>
+
     <!-- Check This Name button -->
+    <!-- *** TODO: keep this as common code at bottom of template -->
     <template v-if="isShowCheckNameButton">
       <v-row justify="center" class="mt-6">
         <v-col cols="auto">
@@ -331,112 +158,50 @@
 </template>
 
 <script lang="ts">
-import { Component, Mixins, Vue, Watch } from 'vue-property-decorator'
-import { Action, Getter } from 'vuex-class'
+import { Component, Mixins } from 'vue-property-decorator'
 
-// Components
 import NameInput from './name-input.vue'
-import NestedSelect from '../common/nested-select.vue'
-import BusinessLookup from '@/components/new-request/business-lookup.vue'
-import BusinessFetch from '@/components/new-request/business-fetch.vue'
 
-// Interfaces / Enums / List Data
-import { BusinessSearchIF, ConversionTypesI, EntityI, FormType, RequestActionsI } from '@/interfaces'
-import { ActionBindingIF } from '@/interfaces/store-interfaces'
-import { AccountType, CompanyTypes, CorpTypeCd, EntityStates, EntityTypes,
-  Location, NrRequestActionCodes, NrRequestTypeCodes } from '@/enums'
-import { CommonMixin, NrAffiliationMixin } from '@/mixins'
-import { BcMapping, CanJurisdictions, ConversionTypes, Designations,
-  IntlJurisdictions, RequestActions, XproMapping } from '@/list-data'
+import BusinessLookupFetch from '@/components/new-request/search-components/business-lookup-fetch.vue'
+import CompanyType from '@/components/new-request/search-components/company-type.vue'
+import CorpNumberCheckbox from '@/components/new-request/search-components/corp-number-checkbox.vue'
+import Designation from '@/components/new-request/search-components/designation.vue'
+import DummyInputBox from '@/components/new-request/search-components/dummy-input-box.vue'
+import EntityType from '@/components/new-request/search-components/entity-type.vue'
+import Jurisdiction from '@/components/new-request/search-components/jurisdiction.vue'
+import NumberedCompanyBullets from '@/components/new-request/search-components/numbered-company-bullets.vue'
+import RequestAction from '@/components/new-request/search-components/request-action.vue'
+import XproFederalBullets from '@/components/new-request/search-components/xpro-federal-bullets.vue'
+
+import { CompanyTypes, EntityTypes } from '@/enums'
+import { CommonMixin, NrAffiliationMixin, SearchMixin } from '@/mixins'
+import { Designations } from '@/list-data'
 import { GetFeatureFlag, Navigate } from '@/plugins'
-import { SessionStorageKeys } from 'sbc-common-components/src/util/constants'
 
 /**
  * This is the component that displays the new NR menus and flows.
  */
 @Component({
-  components: { BusinessLookup, BusinessFetch, NameInput, NestedSelect }
-})
-export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
-  // Refs
-  $refs!: {
-    businessLookupTextFieldForm: FormType
-    selectBusinessTypeRef: FormType
+  components: {
+    BusinessLookupFetch,
+    CompanyType,
+    CorpNumberCheckbox,
+    Designation,
+    DummyInputBox,
+    EntityType,
+    Jurisdiction,
+    NameInput,
+    NumberedCompanyBullets,
+    RequestAction,
+    XproFederalBullets
   }
-
-  // Enums for template
-  readonly CompanyTypes = CompanyTypes
-  readonly NrRequestActionCodes = NrRequestActionCodes
-  readonly RequestActions = RequestActions
-
-  // Store getters
-  @Getter getConversionType!: NrRequestTypeCodes
-  @Getter getConversionTypeOptions!: ConversionTypesI[]
-  @Getter getDesignation!: string
-  @Getter getDisplayedComponent!: string
-  @Getter getEntityBlurbs!: Array<EntityI>
-  @Getter getEntityTypeCd!: EntityTypes
-  @Getter getEntityTypeOptions!: Array<EntityI>
-  @Getter getEntityTextFromValue!: string
-  @Getter getErrors!: string[]
-  @Getter getHasNoCorpNum!: boolean
-  @Getter getLocation!: Location
-  @Getter getOriginEntityTypeCd!: EntityTypes
-  @Getter getRequestActionCd!: NrRequestActionCodes
-  @Getter isAmalgamation!: boolean
-  @Getter isAssumed!: boolean
-  @Getter isAuthenticated!: boolean
-  @Getter isCanadian!: boolean
-  @Getter isChangeName!: boolean
-  @Getter isContinuationIn!: boolean
-  @Getter isConversion!: boolean
-  @Getter isFederal!: boolean
-  @Getter isMobile!: boolean
-  @Getter isMrasJurisdiction!: boolean
-  @Getter isNumberedEntityType!: boolean
-  @Getter isRestoration!: boolean
-  @Getter isXproFlow!: boolean
-
-  // Store actions
-  @Action setConversionType!: ActionBindingIF
-  @Action setCorpNum!: ActionBindingIF
-  @Action setCorpSearch!: ActionBindingIF
-  @Action setClearErrors!: () => void
-  @Action setDesignation!: ActionBindingIF
-  @Action setDoNameCheck!: ActionBindingIF
-  @Action setEntityTypeCd!: ActionBindingIF
-  @Action setExtendedRequestType!: ActionBindingIF
-  @Action setJurisdictionCd!: ActionBindingIF
-  @Action setLocation!: ActionBindingIF
-  @Action setName!: ActionBindingIF
-  @Action setNoCorpNum!: ActionBindingIF
-  @Action setOriginEntityTypeCd!: ActionBindingIF
-  @Action setPickEntityModalVisible!: ActionBindingIF
-  @Action setRequestAction!: ActionBindingIF
-  @Action startAnalyzeName!: ActionBindingIF
-
+})
+export default class Search extends Mixins(CommonMixin, NrAffiliationMixin, SearchMixin) {
   // Constants
   readonly colinLink = sessionStorage.getItem('CORPORATE_ONLINE_URL')
-  readonly corpOnlineLink = 'https://www.corporateonline.gov.bc.ca/'
-  readonly EntityTypes = EntityTypes
-  readonly federalAmalgamationFormLink = 'https://www2.gov.bc.ca/assets/gov/employment-business-and-economic-' +
-    'development/business-management/permits-licences-and-registration/registries-packages/' +
-    'pack_34_xco_-_amalgamation_application_and_business_number_request.pdf'
-  readonly fullReinstatementFormLink = 'https://www2.gov.bc.ca/assets/gov/employment-business-and-economic-' +
-    'development/business-management/permits-licences-and-registration/registries-forms/' +
-    'form_31_xco_-_full_reinstatement_application.pdf'
-  readonly limitedReinstatementFormLink = 'https://www2.gov.bc.ca/assets/gov/employment-business-and-economic-' +
-    'development/business-management/permits-licences-and-registration/registries-forms/' +
-    'form_29_xco_-_limited_reinstatement_application.pdf'
 
-  // Local properties
+  // Variables
   corpNumValid = true
-  request = null as RequestActionsI
-  jurisdiction = null as any
-  selectedCompanyType = null as CompanyTypes
-  activeActionGroup = NaN
-  showRequestActionTooltip = false
-  business = null as BusinessSearchIF
 
   private mounted () {
     this.$nextTick(() => {
@@ -453,67 +218,6 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
     this.scrollTo('namerequest-sbc-header')
   }
 
-  // Text field rules for the business lookup
-  get rules (): any {
-    return [
-      value => {
-        if (this.isConversion) {
-          return this.isPassAlterValidation || 'This business cannot alter their business type'
-        } else if (this.isRestoration) {
-          return this.isRestorable || 'This business cannot be restored.'
-        }
-        return true
-      }]
-  }
-
-  get businessLookupLabel (): string {
-    if (this.isAuthenticated) {
-      if (this.lookupActiveOrHistorical === EntityStates.HISTORICAL) {
-        return 'Find a historical business'
-      } else {
-        return 'Find an existing business'
-      }
-    } else {
-      if (this.lookupActiveOrHistorical === EntityStates.HISTORICAL) {
-        return 'Fetch a historical business'
-      } else {
-        return 'Fetch an existing business'
-      }
-    }
-  }
-
-  get businessLookupHint (): string {
-    if (this.isAuthenticated) {
-      return 'Search by name, incorporation or registration number of existing business'
-    } else {
-      return 'Enter registration number of existing business'
-    }
-  }
-
-  get isBenBusiness (): boolean {
-    return this.business?.legalType as string === CorpTypeCd.BENEFIT_COMPANY
-  }
-
-  /** Search businesses with business lookup depending on the Action selected */
-  get lookupActiveOrHistorical (): String {
-    return this.isRestoration ? EntityStates.HISTORICAL : EntityStates.ACTIVE
-  }
-
-  get isNewBcBusiness (): boolean {
-    return (this.getRequestActionCd === NrRequestActionCodes.NEW_BUSINESS && this.request?.group === 0)
-  }
-
-  get isNewXproBusiness (): boolean {
-    return (this.getRequestActionCd === NrRequestActionCodes.NEW_BUSINESS && this.request?.group === 1)
-  }
-
-  get isPassAlterValidation (): boolean {
-    const legalType = this.business.legalType as unknown as CorpTypeCd
-    return legalType === CorpTypeCd.BC_COMPANY ||
-      legalType === CorpTypeCd.BENEFIT_COMPANY ||
-      legalType === CorpTypeCd.BC_ULC_COMPANY
-  }
-
   get showJurisdiction (): boolean {
     if (this.isAmalgamation && this.isXproEntityType(this.getEntityTypeCd)) return true
     if (this.isNewXproBusiness) return true
@@ -524,7 +228,7 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
 
   get showEntityType (): boolean {
     if (this.isConversion) {
-      return this.business && this.isPassAlterValidation
+      return (!!this.getSearchBusiness && this.isAlterable)
     }
     if (this.getLocation && !this.isFederal && !this.isRestoration && !this.isChangeName) return true
     return false
@@ -535,100 +239,23 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
     return false
   }
 
-  /** Whether current account type is Premium, SBC Staff or Staff. */
-  get isPremiumOrStaff (): boolean {
-    return [AccountType.PREMIUM, AccountType.SBC_STAFF, AccountType.STAFF]
-      .includes(JSON.parse(sessionStorage.getItem(SessionStorageKeys.CurrentAccount))?.accountType)
-  }
-
   get showDesignation (): boolean {
-    if (this.getEntityTypeCd) return (Designations[this.getEntityTypeCd]?.end && !this.isXproFlow)
-    // hide until entity type is selected and needs it
-    return false
-  }
-
-  get designationOptions (): Array<string> {
-    let output = Designations[this.getEntityTypeCd]?.words as string[]
-    if (this.getEntityTypeCd === EntityTypes.CC) {
-      output = Designations[EntityTypes.CR].words
-    }
-    return output
+    // return (Designations[this.getEntityTypeCd]?.end && !this.isXproFlow) // *** TODO: delete
+    return (Designations[this.getEntityTypeCd]?.end || false)
   }
 
   get showCompanyTypeRadioButtons (): boolean {
     if (!this.entity_type_cd && !this.getConversionType && !this.isFederal) return false
     if (this.isConversion) {
-      return this.business !== null && this.getConversionType !== null
+      return (!!this.getSearchBusiness && !!this.getConversionType)
     }
     const isSociety = (this.isSocietyEnabled() && this.getEntityTypeCd === EntityTypes.SO)
     // check if numbered is not allowed or society NR name is required
     if (!this.isNumberedEntityType || isSociety) {
-      this.selectedCompanyType = CompanyTypes.NAMED_COMPANY
+      this.setSearchCompanyType(CompanyTypes.NAMED_COMPANY)
       return false
     }
     return true
-  }
-
-  get isScreenLg () {
-    return this.$vuetify.breakpoint.lgAndUp
-  }
-
-  // FUTURE: clean up return type
-  entityBlurbs (entity_type_cd: string): string | string[] | string[][] {
-    return this.getEntityBlurbs.find(type => type.value === entity_type_cd)?.blurbs
-  }
-
-  get entity_type_cd (): EntityTypes {
-    return this.getEntityTypeCd
-  }
-
-  set entity_type_cd (type: EntityTypes) {
-    // special case for sub-menu
-    if (type === EntityTypes.INFO) {
-      // set empty value until user chooses a new one
-      // (don't use null in case it's already null as we want reactivity)
-      this.setEntityTypeCd('')
-      // show the "View all business types" modal
-      this.setPickEntityModalVisible(true)
-      return
-    }
-    // special case for conversion
-    if (this.getEntityTypeCd && this.isConversion && type) {
-      const value = type as unknown as NrRequestTypeCodes
-      const entity_type_cd = ConversionTypes.find(conv => conv.value === value)?.entity_type_cd || null
-      this.setEntityTypeCd(entity_type_cd)
-      this.setConversionType(type)
-      this.setConversionType(type)
-      return
-    }
-    this.setEntityTypeCd(type)
-  }
-
-  get noCorpNum (): boolean {
-    return this.getHasNoCorpNum
-  }
-
-  set noCorpNum (value: boolean) {
-    this.setNoCorpNum(value)
-  }
-
-  get entityTypeOptions () {
-    return (this.isConversion ? this.getConversionTypeOptions : this.getEntityTypeOptions)
-  }
-
-  get entityConversionText () {
-    // convert NrRequestTypeCodes -> EntityTypes
-    return ConversionTypes.find(conversion => conversion.value === this.getConversionType)?.text
-  }
-
-  /** Whether selected radio button is Named Company. */
-  get isNamedCompany (): boolean {
-    return (this.selectedCompanyType === CompanyTypes.NAMED_COMPANY)
-  }
-
-  /** Whether selected radio button is Numbered Company. */
-  get isNumberedCompany (): boolean {
-    return (this.selectedCompanyType === CompanyTypes.NUMBERED_COMPANY)
   }
 
   get showActionButton (): boolean {
@@ -690,132 +317,9 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
     return 'Incorporate Now'
   }
 
-  get jurisdictionOptions (): Array<any> {
-    const array = []
-
-    // add in Canadian jurisdictions (not including BC)
-    array.push({ isHeader: true, group: 0, text: 'Canadian' })
-    CanJurisdictions
-      .filter(jur => jur.value !== Location.BC)
-      .forEach(jur => array.push({
-        group: 0,
-        text: jur.text,
-        value: jur.value,
-        separator: (jur.value === Location.FD)
-      }))
-
-    // add in International jurisdictions (not including CA)
-    array.push({ isHeader: true, group: 1, text: 'International' })
-    IntlJurisdictions
-      .filter(jur => jur.value !== Location.CA)
-      .forEach(jur => array.push({
-        group: 1,
-        text: jur.text,
-        value: jur.value,
-        separator: false
-      }))
-
-    return array
-  }
-
-  get entityTextFromValue (): string {
-    return this.getEntityTextFromValue || 'specified business type'
-  }
-
-  /** Event handled for business lookup/fetch. */
-  async onBusiness (business: BusinessSearchIF): Promise<void> {
-    this.business = business
-    this.entity_type_cd = this.business?.legalType || null
-    this.setCorpNum(business?.identifier || null)
-    this.setEntityTypeCd(this.business?.legalType)
-    this.setName('')
-
-    // Waiting for DOM update to be able to access the Ref. Trigger form validation.
-    // Need to do that because the ref is in a conditional.
-    await Vue.nextTick()
-    if (this.business) this.$refs.businessLookupTextFieldForm.validate()
-
-    if (this.isConversion) {
-      if (this.business) {
-        // set the from business for alteration (conversion)
-        this.setOriginEntityTypeCd(this.business.legalType)
-        // special case if the from business is BEN
-        // set conversionType & entity_type_cd because there's only one alteration type for it
-        if (this.isBenBusiness) {
-          this.setConversionType(NrRequestTypeCodes.CONVERT_CORP)
-          this.setEntityTypeCd(EntityTypes.BC)
-        }
-      } else {
-        // clear all related fields when clearing business search/fetch for alter
-        this.setConversionType(null)
-        this.setOriginEntityTypeCd(null)
-        this.selectedCompanyType = null
-      }
-    }
-
-    if (this.isRestoration) {
-      // Check if not XPRO and BC restorable
-      if (!this.isSelectedXproAndRestorable && this.isBcRestorable) {
-        this.setLocation(Location.BC)
-        this.setEntityTypeCd(this.corpTypeToEntityType(this.business.legalType as unknown as CorpTypeCd))
-      } else if (this.isSelectedXproAndRestorable) { // Check if XPRO and restorable
-        this.setLocation(Location.CA)
-        this.setEntityTypeCd(this.business.legalType)
-      } else {
-        this.setEntityTypeCd(null)
-      }
-    }
-
-    if (this.isChangeName) {
-      if (this.isBenBcUlcCccCorpType) {
-        this.setLocation(Location.BC)
-        this.setEntityTypeCd(this.corpTypeToEntityType(this.business?.legalType as unknown as CorpTypeCd))
-      } else {
-        this.setEntityTypeCd(this.business.legalType)
-      }
-
-      if (this.isChangeNameXpro) {
-        this.setLocation(Location.CA)
-        this.setEntityTypeCd(this.business.legalType)
-      }
-    }
-  }
-
-  /** Whether the selected XPRO is restorable. */
-  get isSelectedXproAndRestorable (): boolean {
-    return XproMapping.REH.includes(this.business?.legalType)
-  }
-
-  /** Whether the selected business' legal type is BC and restorable. */
-  get isBcRestorable (): boolean {
-    return BcMapping.REH.includes(this.corpTypeToEntityType(this.business?.legalType as unknown as CorpTypeCd))
-  }
-
-  /** Whether company is restorable. */
-  get isRestorable (): boolean {
-    return this.isSelectedXproAndRestorable || this.isBcRestorable
-  }
-
-  /** Map to correct Entity type if legal type is Ben Bc Ulc Ccc */
-  isBenBcUlcCccCorpType (): boolean {
-    return (
-      this.business?.legalType === EntityTypes.BC ||
-      this.business?.legalType === EntityTypes.BEN ||
-      this.business?.legalType === EntityTypes.CC ||
-      this.business?.legalType === EntityTypes.ULC
-    )
-  }
-
-  get isChangeNameXpro (): boolean {
-    return XproMapping.CHG.includes(this.business?.legalType)
-  }
-
-  get isSelectedCompanyXPro (): boolean {
-    if (this.isRestoration) {
-      return this.isSelectedXproAndRestorable
-    }
-    if (this.isChangeName) {
-      return this.isChangeNameXpro
+  get isShowCheckNameButton (): boolean {
+    if (!this.isFederal && this.isNamedCompany && this.entity_type_cd) {
+      return true
     }
     return false
   }
@@ -829,7 +333,7 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
     const legalType = this.entityTypeToCorpType(this.entity_type_cd)
     if (this.isAuthenticated) {
       if (this.isConversion || this.isRestoration || this.isChangeName) {
-        this.goToEntityDashboard(this.business.identifier)
+        this.goToEntityDashboard(this.getSearchBusiness.identifier)
       } else {
         await this.incorporateNow(legalType)
       }
@@ -849,103 +353,11 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
     await this.startAnalyzeName(null)
     if (this.isXproFlow) this.$root.$emit('showSpinner', false)
   }
-
-  get isShowCheckNameButton (): boolean {
-    if (!this.isFederal && this.isNamedCompany && this.entity_type_cd) {
-      return true
-    }
-    return false
-  }
-
-  @Watch('entity_type_cd')
-  clearDesignation () {
-    this.setDesignation('')
-    // clear "Select a Business Type" field when "View all business types" or Society is selected
-    if (!this.entity_type_cd || this.entity_type_cd === EntityTypes.INFO) {
-      this.$refs.selectBusinessTypeRef && this.$refs.selectBusinessTypeRef.reset()
-    }
-  }
-
-  /** Resets search values when location changes. */
-  @Watch('getLocation')
-  watchLocation (newVal: Location) {
-    // if they need to search by corp num first then reset the name
-    if ([Location.CA, Location.FD, Location.IN, Location.US].includes(newVal)) {
-      this.setName('')
-    }
-    this.setCorpSearch('')
-    this.setNoCorpNum(false)
-  }
-
-  /** Resets fields when returned to the Tabs component */
-  @Watch('getDisplayedComponent')
-  watchDisplayedComponent (displayedComponent: string) {
-    if (displayedComponent === 'Tabs') {
-      this.onBusiness(null)
-    }
-  }
-
-  /** Called when Request Action menu item is changed. */
-  async onRequestActionChange (request: RequestActionsI): Promise<void> {
-    this.request = request
-
-    this.setRequestAction(this.request?.value || null)
-
-    // clear previous state
-    this.business = null
-    this.jurisdiction = null
-    this.setLocation(null)
-    this.setJurisdictionCd(null)
-    if (this.entity_type_cd) this.entity_type_cd = null
-    this.selectedCompanyType = null
-    this.setCorpNum(null)
-    this.setName('')
-
-    // wait for updates
-    await Vue.nextTick()
-
-    if (this.getRequestActionCd !== NrRequestActionCodes.NEW_BUSINESS) {
-      this.setExtendedRequestType(this.request)
-    }
-
-    // set default location for requests where there is only one location option
-    if (this.isNewBcBusiness || this.isContinuationIn || this.isConversion || this.isAmalgamation) {
-      // *** TODO: set location for amalgamation later (depending on entity type selected)
-      this.setLocation(Location.BC)
-    } else if (this.isAssumed && this.getLocation === Location.BC) {
-      this.setLocation(Location.CA)
-    }
-
-    // calculate whether to show tooltip
-    // (after waiting for DOM update)
-    await Vue.nextTick()
-    const el = document.querySelector('.request-action-select .v-select__selection') as any
-    const offsetWidth = el?.offsetWidth as number
-    const scrollWidth = el?.scrollWidth as number
-    this.showRequestActionTooltip = (offsetWidth < scrollWidth)
-  }
-
-  /** Called when Jurisdiction menu item is changed. */
-  onJurisdictionChange (jurisdiction: any): void {
-    this.jurisdiction = jurisdiction
-
-    this.setLocation(jurisdiction.group === 0 ? Location.CA : Location.IN)
-    this.setJurisdictionCd(jurisdiction.value)
-    // Resetting the entity type when a business is selected (after jurisdiction change)
-    if (this.business) {
-      this.setEntityTypeCd(this.business.legalType)
-    }
-  }
 }
 </script>
 
 <style lang="scss" scoped>
 @import '@/assets/styles/theme.scss';
-
-.v-icon.mdi-phone,
-.v-icon.mdi-email {
-  color: $app-dk-blue;
-}
 
 .v-list {
   padding: 0;
@@ -954,21 +366,6 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
 .list-item {
   width: 100%;
   padding: 8px;
-}
-
-.last-select-item {
-  border-top: 1px solid $gray3;
-  padding: 20px 8px !important;
-}
-
-// hide disabled list items
-::v-deep .v-list-item:has(.v-list-item__content.hide-me) {
-  display: none;
-}
-
-// make the business lookup close button always blue
-::v-deep .business-lookup button.mdi-close {
-  color: $app-blue;
 }
 
 // set content colour when hovering over list items
@@ -1005,12 +402,6 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
   width: 17rem !important;
 }
 
-// Line spacing between bullet points and sizing
-.bullet-points {
-  line-height: 1.5rem;
-  font-size: $px-15;
-}
-
 // Vuetify overrides
 ::v-deep {
   .theme--light.v-btn.v-btn--disabled:not(.v-btn--flat):not(.v-btn--text):not(.v-btn--outlined) {
@@ -1032,11 +423,6 @@ export default class Search extends Mixins(CommonMixin, NrAffiliationMixin) {
 
   .v-select__selection--disabled {
     color: $gray9 !important;
-  }
-
-  // reduce checkbox height when there are no error messages
-  .v-messages:not(.error--text) {
-    margin-bottom: -22px;
   }
 
   .theme--light.v-list-item:not(.v-list-item--active):not(.v-list-item--disabled) {
