+++ src/mixins/index.ts
@@ -6,4 +6,5 @@ export * from './date-mixin'
 export * from './displayed-component-mixin'
 export * from './load-keycloak-roles-mixin'
 export * from './nr-affiliation-mixin'
+export * from './search-mixin'
 export * from './update-user-mixin'
